+++ packages/pmml-editor/src/editor/index.ts
@@ -19,3 +19,4 @@ import "@patternfly/patternfly/patternfly.css";
 export * from "./PMMLEditor";
 export * from "./PMMLEditorFactory";
 export * from "./PMMLEditorInterface";
+export * from "./PMMLModelHelper";
